Change the screenshot upload retry logic
The uploads need to be retried when the save operation fails, because
if any one of the screenshot uploads fail, you don't find out until the
save operation.

I set things up so that we save after every device size is uploaded, and
then repeat the uploads if the save operation fails. In a perfect world,
I would prefer to not allow the save operation to retry at all, and
instead start re-uploading after the first exception. But…this works.